Move most of the ECS related code to the plugin
We're about to add an infrastructure type module which will require
different behaviour from release/deploy/destroy so we've pluginised
the code into the ecs plugin module.

JIRA: PLAT-834
