+++ cdflow_commands/cli.py
@@ -29,14 +29,13 @@ from cdflow_commands.config import (
     get_component_name,
     get_platform_config_path
 )
-from cdflow_commands.release import Release, ReleaseConfig
-from cdflow_commands.deploy import Deploy, DeployConfig
-from cdflow_commands.destroy import Destroy
-from cdflow_commands.ecs_monitor import ECSEventIterator, ECSMonitor
 from cdflow_commands.exceptions import UserFacingError
 from cdflow_commands.logger import logger
 from cdflow_commands.terragrunt import S3BucketFactory, write_terragrunt_config
-from cdflow_commands.plugins.ecs import ECSPlugin
+from cdflow_commands.plugins.ecs import (
+    Deploy, DeployConfig, ECSPlugin, Release, ReleaseConfig, Destroy,
+    ECSEventIterator, ECSMonitor
+)
 
 
 def run(argv):
