+++ test/plugins/test_ecs_deploy.py
@@ -7,7 +7,7 @@ from hypothesis.strategies import text, fixed_dictionaries, dictionaries
 from boto3 import Session
 from mock import ANY
 
-from cdflow_commands.deploy import Deploy, DeployConfig
+from cdflow_commands.plugins.ecs import Deploy, DeployConfig
 
 
 IGNORED_PARAMS = [ANY] * 18
@@ -30,9 +30,9 @@ class TestDeploy(unittest.TestCase):
             'dummy-version', 'dummy-ecs-cluster', self._deploy_config
         )
 
-    @patch('cdflow_commands.deploy.check_call')
-    @patch('cdflow_commands.deploy.get_secrets')
-    @patch('cdflow_commands.deploy.NamedTemporaryFile')
+    @patch('cdflow_commands.plugins.ecs.check_call')
+    @patch('cdflow_commands.plugins.ecs.get_secrets')
+    @patch('cdflow_commands.plugins.ecs.NamedTemporaryFile')
     def test_terraform_modules_fetched(
         self, NamedTemporaryFile, get_secrets, check_call
     ):
@@ -43,9 +43,9 @@ class TestDeploy(unittest.TestCase):
         # Then
         check_call.assert_any_call(['terragrunt', 'get', 'infra'])
 
-    @patch('cdflow_commands.deploy.check_call')
-    @patch('cdflow_commands.deploy.get_secrets')
-    @patch('cdflow_commands.deploy.NamedTemporaryFile')
+    @patch('cdflow_commands.plugins.ecs.check_call')
+    @patch('cdflow_commands.plugins.ecs.get_secrets')
+    @patch('cdflow_commands.plugins.ecs.NamedTemporaryFile')
     def test_terragrunt_plan_called(
         self, NamedTemporaryFile, get_secrets, check_call
     ):
@@ -59,9 +59,9 @@ class TestDeploy(unittest.TestCase):
             env=ANY
         )
 
-    @patch('cdflow_commands.deploy.check_call')
-    @patch('cdflow_commands.deploy.get_secrets')
-    @patch('cdflow_commands.deploy.NamedTemporaryFile')
+    @patch('cdflow_commands.plugins.ecs.check_call')
+    @patch('cdflow_commands.plugins.ecs.get_secrets')
+    @patch('cdflow_commands.plugins.ecs.NamedTemporaryFile')
     def test_terrgrunt_apply_called(
         self, NamedTemporaryFile, get_secrets, check_call
     ):
@@ -97,11 +97,11 @@ class TestDeploy(unittest.TestCase):
         deploy = Deploy(boto_session, ANY, ANY, ANY, ANY, self._deploy_config)
 
         with patch(
-            'cdflow_commands.deploy.check_call'
+            'cdflow_commands.plugins.ecs.check_call'
         ) as check_call, patch(
-            'cdflow_commands.deploy.NamedTemporaryFile', autospec=True
+            'cdflow_commands.plugins.ecs.NamedTemporaryFile', autospec=True
         ) as NamedTemporaryFile, patch(
-            'cdflow_commands.deploy.get_secrets'
+            'cdflow_commands.plugins.ecs.get_secrets'
         ) as get_secrets:
             NamedTemporaryFile.return_value.__enter__.return_value.name = ANY
             get_secrets.return_value = {}
@@ -151,13 +151,13 @@ class TestDeploy(unittest.TestCase):
         deploy = Deploy(boto_session, ANY, ANY, ANY, ANY, self._deploy_config)
 
         with patch(
-            'cdflow_commands.deploy.os'
+            'cdflow_commands.plugins.ecs.os'
         ) as mock_os, patch(
-            'cdflow_commands.deploy.check_call'
+            'cdflow_commands.plugins.ecs.check_call'
         ) as check_call, patch(
-            'cdflow_commands.deploy.NamedTemporaryFile', autospec=True
+            'cdflow_commands.plugins.ecs.NamedTemporaryFile', autospec=True
         ) as NamedTemporaryFile, patch(
-            'cdflow_commands.deploy.get_secrets'
+            'cdflow_commands.plugins.ecs.get_secrets'
         ) as get_secrets:
             NamedTemporaryFile.return_value.__enter__.return_value.name = ANY
             get_secrets.return_value = {}
@@ -198,13 +198,13 @@ class TestDeploy(unittest.TestCase):
         deploy = Deploy(boto_session, ANY, ANY, ANY, ANY, self._deploy_config)
 
         with patch(
-            'cdflow_commands.deploy.os'
+            'cdflow_commands.plugins.ecs.os'
         ) as mock_os, patch(
-            'cdflow_commands.deploy.check_call'
+            'cdflow_commands.plugins.ecs.check_call'
         ), patch(
-            'cdflow_commands.deploy.NamedTemporaryFile', autospec=True
+            'cdflow_commands.plugins.ecs.NamedTemporaryFile', autospec=True
         ) as NamedTemporaryFile, patch(
-            'cdflow_commands.deploy.get_secrets'
+            'cdflow_commands.plugins.ecs.get_secrets'
         ) as get_secrets:
             NamedTemporaryFile.return_value.__enter__.return_value.name = ANY
             get_secrets.return_value = {}
@@ -260,11 +260,11 @@ class TestDeploy(unittest.TestCase):
 
         # When
         with patch(
-            'cdflow_commands.deploy.check_call'
+            'cdflow_commands.plugins.ecs.check_call'
         ) as check_call, patch(
-            'cdflow_commands.deploy.NamedTemporaryFile', autospec=True
+            'cdflow_commands.plugins.ecs.NamedTemporaryFile', autospec=True
         ) as NamedTemporaryFile, patch(
-            'cdflow_commands.deploy.get_secrets'
+            'cdflow_commands.plugins.ecs.get_secrets'
         ) as get_secrets:
             NamedTemporaryFile.return_value.__enter__.return_value.name = \
                 secret_file_path
@@ -315,13 +315,13 @@ class TestEnvironmentSpecificConfigAddedToTerraformArgs(unittest.TestCase):
 
         # When
         with patch(
-            'cdflow_commands.deploy.check_call'
+            'cdflow_commands.plugins.ecs.check_call'
         ) as check_call, patch(
-            'cdflow_commands.deploy.path'
+            'cdflow_commands.plugins.ecs.path'
         ) as path, patch(
-            'cdflow_commands.deploy.NamedTemporaryFile', autospec=True
+            'cdflow_commands.plugins.ecs.NamedTemporaryFile', autospec=True
         ) as NamedTemporaryFile, patch(
-            'cdflow_commands.deploy.get_secrets'
+            'cdflow_commands.plugins.ecs.get_secrets'
         ) as get_secrets:
             NamedTemporaryFile.return_value.__enter__.return_value.name = ANY
             get_secrets.return_value = {}
