+++ test/plugins/test_ecs_destroy.py
@@ -7,7 +7,7 @@ from mock import patch, ANY, Mock
 from hypothesis import given
 from hypothesis.strategies import text, dictionaries, fixed_dictionaries
 
-from cdflow_commands.destroy import Destroy
+from cdflow_commands.plugins.ecs import Destroy
 
 BotoCreds = namedtuple('BotoCreds', ['access_key', 'secret_key', 'token'])
 
@@ -31,7 +31,7 @@ class TestDestroy(unittest.TestCase):
             boto_session, component_name, environment_name, 'dummy-bucket'
         )
 
-        with patch('cdflow_commands.destroy.check_call') as check_call:
+        with patch('cdflow_commands.plugins.ecs.check_call') as check_call:
             destroy.run()
             check_call.assert_any_call([
                 'terragrunt', 'plan', '-destroy',
@@ -53,7 +53,7 @@ class TestDestroy(unittest.TestCase):
             boto_session, component_name, environment_name, 'dummy-bucket'
         )
 
-        with patch('cdflow_commands.destroy.check_call') as check_call:
+        with patch('cdflow_commands.plugins.ecs.check_call') as check_call:
             destroy.run()
             check_call.assert_any_call([
                 'terragrunt', 'destroy', '-force',
@@ -79,7 +79,7 @@ class TestDestroy(unittest.TestCase):
             'dummy-environment', 'dummy-bucket'
         )
 
-        with patch('cdflow_commands.destroy.check_call') as check_call:
+        with patch('cdflow_commands.plugins.ecs.check_call') as check_call:
             destroy.run()
 
             env = check_call.mock_calls[0][CALL_KWARGS]['env']
@@ -111,9 +111,9 @@ class TestDestroy(unittest.TestCase):
         )
 
         with patch(
-            'cdflow_commands.destroy.check_call'
+            'cdflow_commands.plugins.ecs.check_call'
         ) as check_call, patch(
-            'cdflow_commands.destroy.os'
+            'cdflow_commands.plugins.ecs.os'
         ) as mock_os:
             mock_os.environ = mock_env.copy()
             destroy.run()
@@ -150,7 +150,7 @@ class TestDestroy(unittest.TestCase):
         )
 
         # When
-        with patch('cdflow_commands.destroy.check_call'):
+        with patch('cdflow_commands.plugins.ecs.check_call'):
             destroy.run()
 
         # Then
